#3: vote into separate app
