+++ swot/serializers.py
@@ -1,15 +1,9 @@
 from rest_framework import serializers
 
-from swot.models import SwotItem, Vote
+from swot.models import SwotItem
 
 
 class SwotItemSerializer(serializers.ModelSerializer):
     class Meta:
         model = SwotItem
         fields = ('id', 'cardType', 'text')
-
-
-class VoteSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Vote
-        fields = ('id', 'item', 'voteType')
