+++ swot_item_vote/tests/test_views_vote.py
